Fjerner unødvendig kode fra oppgave
