+++ apps/etterlatte-behandling/src/main/kotlin/oppgave/OppgaveRoutes.kt
@@ -287,7 +287,3 @@ internal fun Route.oppgaveRoutes(service: OppgaveService) {
         }
     }
 }
-
-internal data class GosysOppgaveversjon(
-    val versjon: Long,
-)
